feat(verifier): add tool to verify digital signature
