+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/tokopedia/cryptography-example
+
+go 1.14
